refactor + linting manager
split into separate package
change ProcessHandler to SourceHandler

update tests
